BREAKING CHANGE: Call provider functions with objects
Previously, privder functions expected that dependencies would be passed
as arguments in the order in which they were specified in the provider
definitition.  This change makes it so that the provider is invoked with
an object with keys matching the names of the dependencies.

This change is being made in order to reduce the chances of bugs due to
mismatched the order of the argument names in the function not matching
the order of the dependency list.  This change is especially useful when
the framework is used in combination with ES6 object destructuring.