Add put method for adding key/promise to slate
In order to support reusing a slate instance in conditions where some of
the inputs change after a call to clear, include a put method to add a
key/promise pair to the slate.  Anything added to the slate through put
will be lost once clear is called.  The expectation is that the workflow
of using slate would involve calling put with invocation specific
values, calling get to resove invocaion specific values, then calling
clear to reset and repeating the cycle with new invocation specific
values.